Move confighandle.migration package to config package.
