+++ src/main/java/org/mvplugins/multiverse/core/config/CoreConfig.java
@@ -19,12 +19,12 @@ import org.mvplugins.multiverse.core.command.MVCommandManager;
 import org.mvplugins.multiverse.core.command.queue.ConfirmMode;
 import org.mvplugins.multiverse.core.config.handle.CommentedConfigurationHandle;
 import org.mvplugins.multiverse.core.config.handle.StringPropertyHandle;
-import org.mvplugins.multiverse.core.confighandle.migration.BooleanMigratorAction;
-import org.mvplugins.multiverse.core.confighandle.migration.ConfigMigrator;
-import org.mvplugins.multiverse.core.confighandle.migration.IntegerMigratorAction;
-import org.mvplugins.multiverse.core.confighandle.migration.InvertBoolMigratorAction;
-import org.mvplugins.multiverse.core.confighandle.migration.MoveMigratorAction;
-import org.mvplugins.multiverse.core.confighandle.migration.VersionMigrator;
+import org.mvplugins.multiverse.core.config.migration.BooleanMigratorAction;
+import org.mvplugins.multiverse.core.config.migration.ConfigMigrator;
+import org.mvplugins.multiverse.core.config.migration.IntegerMigratorAction;
+import org.mvplugins.multiverse.core.config.migration.InvertBoolMigratorAction;
+import org.mvplugins.multiverse.core.config.migration.MoveMigratorAction;
+import org.mvplugins.multiverse.core.config.migration.VersionMigrator;
 import org.mvplugins.multiverse.core.destination.DestinationsProvider;
 
 @Service
