+++ src/main/java/org/mvplugins/multiverse/core/world/WorldConfig.java
@@ -17,16 +17,16 @@ import org.jetbrains.annotations.Nullable;
 import org.mvplugins.multiverse.core.MultiverseCore;
 import org.mvplugins.multiverse.core.config.handle.MemoryConfigurationHandle;
 import org.mvplugins.multiverse.core.config.handle.StringPropertyHandle;
-import org.mvplugins.multiverse.core.confighandle.migration.BooleanMigratorAction;
-import org.mvplugins.multiverse.core.confighandle.migration.ConfigMigrator;
-import org.mvplugins.multiverse.core.confighandle.migration.DeleteMigratorAction;
-import org.mvplugins.multiverse.core.confighandle.migration.DoubleMigratorAction;
-import org.mvplugins.multiverse.core.confighandle.migration.IntegerMigratorAction;
-import org.mvplugins.multiverse.core.confighandle.migration.LongMigratorAction;
-import org.mvplugins.multiverse.core.confighandle.migration.MigratorAction;
-import org.mvplugins.multiverse.core.confighandle.migration.MoveMigratorAction;
-import org.mvplugins.multiverse.core.confighandle.migration.NullStringMigratorAction;
-import org.mvplugins.multiverse.core.confighandle.migration.VersionMigrator;
+import org.mvplugins.multiverse.core.config.migration.BooleanMigratorAction;
+import org.mvplugins.multiverse.core.config.migration.ConfigMigrator;
+import org.mvplugins.multiverse.core.config.migration.DeleteMigratorAction;
+import org.mvplugins.multiverse.core.config.migration.DoubleMigratorAction;
+import org.mvplugins.multiverse.core.config.migration.IntegerMigratorAction;
+import org.mvplugins.multiverse.core.config.migration.LongMigratorAction;
+import org.mvplugins.multiverse.core.config.migration.MigratorAction;
+import org.mvplugins.multiverse.core.config.migration.MoveMigratorAction;
+import org.mvplugins.multiverse.core.config.migration.NullStringMigratorAction;
+import org.mvplugins.multiverse.core.config.migration.VersionMigrator;
 import org.mvplugins.multiverse.core.economy.MVEconomist;
 import org.mvplugins.multiverse.core.world.location.SpawnLocation;
